커뮤니티 crud 댓글 crud

+++ src/main/java/org/nurim/nurim/service/CommunityService.java
@@ -2,35 +2,38 @@ package org.nurim.nurim.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 import org.nurim.nurim.domain.dto.community.*;
 import org.nurim.nurim.domain.dto.reply.ReadReplyResponse;
 import org.nurim.nurim.domain.entity.Community;
-import org.nurim.nurim.domain.entity.CommunityImage;
 import org.nurim.nurim.domain.entity.Member;
-import org.nurim.nurim.domain.entity.Recommend;
-import org.nurim.nurim.repository.CommunityImageRepository;
-import org.nurim.nurim.repository.CommunityRepository;
-import org.nurim.nurim.repository.MemberRepository;
+import org.nurim.nurim.domain.entity.Reply;
+import org.nurim.nurim.repository.*;
 //import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.nurim.nurim.repository.RecommendRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.nio.file.AccessDeniedException;
+import java.util.List;
+import java.util.Objects;
+
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
+@Log4j2
 public class CommunityService {
 
     private final CommunityRepository communityRepository;
 
     private final MemberRepository memberRepository;
 
-    private final CommunityImageRepository communityImageRepository;
+    private final ReplyService replyService;
 
-    private final  RecommendRepository recommendRepository;
-
+    /**
+     * 게시물 생성 서비스
+     */
     @Transactional
     public CreateCommunityResponse communityCreate(Long memberId, CreateCommunityRequest request) {
 
@@ -53,10 +56,16 @@ public class CommunityService {
                 saveCommunity.getCommunityCategory());
     }
 
+    /**
+     * 게시물 단일 조회 및 조회수,댓글출력 서비스
+     */
+
     @Transactional
     public ReadCommunityResponse communityRead(Long communityId) {
         Community findCommunity = communityRepository.findById(communityId)
                 .orElseThrow(() -> new EntityNotFoundException("해당 communityId로 조회된 게시글이 없습니다."));
+        log.info("커뮤니티 조회 성공");
+        List<ReadReplyResponse> listReply = replyService.getRepliesByCommunityId(communityId);
         communityRepository.updateCount(communityId);
         return new ReadCommunityResponse(
                 findCommunity.getCommunityId(),
@@ -66,41 +75,54 @@ public class CommunityService {
                 findCommunity.getCommunityCategory(),
                 findCommunity.getRegisterDate(),
                 findCommunity.getModifyDate(),
-                findCommunity.getCounts(),
-                findCommunity.getRecommend());
+                findCommunity.getViewCounts(),
+                findCommunity.getRecommend(),
+                findCommunity.getMember().getMemberNickname(),
+                listReply);
     }
 
+    /**
+     *게시글 삭제 글쓴이 본인만!
+     */
     @Transactional
-    public DeleteCommunityResponse communityDelete(Long communityId) {
+    public DeleteCommunityResponse communityDelete(Long communityId, Long memberId) throws AccessDeniedException {
         Community findCommunity = communityRepository.findById(communityId)
                 .orElseThrow(() -> new EntityNotFoundException("해당 communityId로 조회된 게시글이 없습니다."));
 
-        communityRepository.delete(findCommunity);
+        Long memberInCommunityId = findCommunity.getMember().getMemberId();
 
-        return new DeleteCommunityResponse(findCommunity.getCommunityId());
+        if (Objects.equals(memberInCommunityId, memberId)) {
+            communityRepository.delete(findCommunity);
+            return new DeleteCommunityResponse(findCommunity.getCommunityId());
+        } else {
+            throw new AccessDeniedException("커뮤니티를 삭제할 권한이 없습니다.");
         }
+    }
+
 
     @Transactional
-    public UpdateCommunityResponse communityUpdate(Long communityId, UpdateCommunityRequest request){
+    public UpdateCommunityResponse communityUpdate(Long communityId, UpdateCommunityRequest request) {
 
         Community findCommunity = communityRepository.findById(communityId)
                 .orElseThrow(() -> new EntityNotFoundException("해당 아이디로 조회안됨"));
 
-        findCommunity.update(request.getTitle(),request.getContent(),request.getCategory());
+        Long memberInCommunityId = findCommunity.getMember().getMemberId();
+
+        findCommunity.update(request.getTitle(), request.getContent(), request.getCategory());
 
         return new UpdateCommunityResponse(
                 findCommunity.getTitle(),
                 findCommunity.getContent(),
-                findCommunity.getModifyDate(),
-                findCommunity.getCommunityCategory());
+                findCommunity.getModifyDate());
     }
-    public Page<ReadCommunityResponse> getCommunityListByCategory(String category, Pageable pageable){
+
+    public Page<ReadSearchResponse> getCommunityListByCategory(String category, Pageable pageable) {
         Page<Community> communityPage = communityRepository.findByCommunityCategory(category, pageable);
 
         return communityPage.map(community -> {
             Long memberId = community.getMember().getMemberId();
 
-            return new ReadCommunityResponse(
+            return new ReadSearchResponse(
                     community.getCommunityId(),
                     community.getCommunityImage(),
                     community.getTitle(),
@@ -108,15 +130,130 @@ public class CommunityService {
                     community.getCommunityCategory(),
                     community.getRegisterDate(),
                     community.getModifyDate(),
-                    community.getCounts(),
-                    community.getRecommend());
+                    community.getViewCounts(),
+                    community.getRecommend(),
+                    community.getMember().getMemberNickname());
         });
     }
 
-    @Transactional
-    public int updateCount(Long communityId){
-        return communityRepository.updateCount(communityId);
+    /**
+     *조회수 많은 게시글 5개
+     */
+    public Page<ReadCountsCommunitiesResponse> findPopularCommunities(Pageable pageable) {
+        Page<Community> popularPage = communityRepository.findAll(pageable);
+
+        return popularPage.map(community -> {
+            Long memberId = community.getMember().getMemberId();
+
+            return new ReadCountsCommunitiesResponse(
+                    community.getCommunityId(),
+                    community.getTitle(),
+                    community.getCommunityCategory(),
+                    community.getViewCounts(),
+                    community.getRecommend(),
+                    community.getMember().getMemberNickname());
+        });
     }
 
-}
+    public Page<ReadSearchResponse> SearchTitleAndCategoryAndMemberNickName(String communityTitle,String communityCategory,String memberNickname, Pageable pageable) {
+        Page<Community> searchPage = communityRepository.findByTitleAndCommunityCategoryAndMember_MemberNickname(communityTitle,communityCategory,memberNickname, pageable);
 
+        return searchPage.map(community -> {
+            Long memberId = community.getMember().getMemberId();
+
+            return new ReadSearchResponse(
+                    community.getCommunityId(),
+                    community.getCommunityImage(),
+                    community.getTitle(),
+                    community.getContent(),
+                    community.getCommunityCategory(),
+                    community.getRegisterDate(),
+                    community.getModifyDate(),
+                    community.getViewCounts(),
+                    community.getRecommend(),
+                    community.getMember().getMemberNickname());
+        });
+
+    }
+    public Page<ReadSearchResponse> SearchTitle(String communityTitle,Pageable pageable) {
+        Page<Community> searchPage = communityRepository.findByTitle(communityTitle,pageable);
+
+        return searchPage.map(community -> {
+            Long memberId = community.getMember().getMemberId();
+
+            return new ReadSearchResponse(
+                    community.getCommunityId(),
+                    community.getCommunityImage(),
+                    community.getTitle(),
+                    community.getContent(),
+                    community.getCommunityCategory(),
+                    community.getRegisterDate(),
+                    community.getModifyDate(),
+                    community.getViewCounts(),
+                    community.getRecommend(),
+                    community.getMember().getMemberNickname());
+        });
+
+    }
+    public Page<ReadSearchResponse> SearchCategory(String communityCategory, Pageable pageable) {
+        Page<Community> searchPage = communityRepository.findByCommunityCategory(communityCategory, pageable);
+
+        return searchPage.map(community -> {
+            Long memberId = community.getMember().getMemberId();
+
+            return new ReadSearchResponse(
+                    community.getCommunityId(),
+                    community.getCommunityImage(),
+                    community.getTitle(),
+                    community.getContent(),
+                    community.getCommunityCategory(),
+                    community.getRegisterDate(),
+                    community.getModifyDate(),
+                    community.getViewCounts(),
+                    community.getRecommend(),
+                    community.getMember().getMemberNickname());
+        });
+
+    }
+
+    public Page<ReadSearchResponse> SearchMemberNickName(String memberNickname, Pageable pageable) {
+        Page<Community> searchPage = communityRepository.findByMemberMemberNickname(memberNickname, pageable);
+
+        return searchPage.map(community -> {
+            Long memberId = community.getMember().getMemberId();
+
+            return new ReadSearchResponse(
+                    community.getCommunityId(),
+                    community.getCommunityImage(),
+                    community.getTitle(),
+                    community.getContent(),
+                    community.getCommunityCategory(),
+                    community.getRegisterDate(),
+                    community.getModifyDate(),
+                    community.getViewCounts(),
+                    community.getRecommend(),
+                    community.getMember().getMemberNickname());
+        });
+
+    }
+//    public Page<ReadSearchResponse> SearchTitleAndMemberNickName(String title,String memberNickname, Pageable pageable) {
+//        Page<Community> searchPage = communityRepository.findByTitleAndMemberMemberNickname(title,memberNickname, pageable);
+//
+//        return searchPage.map(community -> {
+//            Long memberId = community.getMember().getMemberId();
+//
+//            return new ReadSearchResponse(
+//                    community.getCommunityId(),
+//                    community.getCommunityImage(),
+//                    community.getTitle(),
+//                    community.getContent(),
+//                    community.getCommunityCategory(),
+//                    community.getRegisterDate(),
+//                    community.getModifyDate(),
+//                    community.getViewCounts(),
+//                    community.getRecommend(),
+//                    community.getMember().getMemberNickname());
+//        });
+//
+//    }
+}
